Set up GoReleaser to make releases for tags
Related: #17

Adding [GoReleaser](https://goreleaser.com/) to handle generating
releases for tags. This will make it so that cutting releases is as
simple as pushing a tag. The result will be a release which includes the
changelog, binaries for Linux/Windows/Darwin, as well as the checksums
for each of the archives.

I've added some light details about how it works in RELEASE.md as well
as the one-time setup step required to make it work (A limited scope
GitHub token).